refactor(sqlite): Move error types into new error module
